# Optimize expression compilation memory usage