Fixes: https://github.com/aspnet/DependencyInjection/issues/425 by caching private field access into local variables.
Also optimizes amount of `Monitor.Enter/Exit` and `try/catch` we do for scope services.

## Test case:
IdentitySample.Mvc project
Run `net451` and request `/` multiple times to force all services to compile.

## Before
**234 MB** allocation from `Microsoft.Extensions.DependencyInjection.ServiceProvider+<>c__DisplayClass17_0.<RealizeService>b__1()`, **223 MB** because of CAS checks.
And **2700ms** total compilation time.
![http://i.imgur.com/mm7VunG.png](http://i.imgur.com/mm7VunG.png)

## After
**9.8 MB** allocation from `Microsoft.Extensions.DependencyInjection.ServiceProvider+<>c__DisplayClass17_0.<RealizeService>b__1()`, **4.3 MB** because of CAS checks.
And **400ms** total compilation time.

![http://i.imgur.com/pJkQSj5.png](http://i.imgur.com/pJkQSj5.png)

## After + No trees for singletons
**2.7 MB** allocation from `Microsoft.Extensions.DependencyInjection.ServiceProvider+<>c__DisplayClass17_0.<RealizeService>b__1()`, **1.8 MB** because of CAS checks.
And **122ms** total compilation time.

![http://i.imgur.com/xIep1KC.png](http://i.imgur.com/xIep1KC.png)

## After + No trees for singletons + Short circuit singletons
https://github.com/aspnet/DependencyInjection/pull/429/files#diff-800d14c6aa47e810955d46f860d9502cR52

**1.6 MB** allocation from `Microsoft.Extensions.DependencyInjection.ServiceLookup.CallSiteExpressionBuilder.Build(IServiceCallSite)`, **1.2 MB** because of CAS checks.
And **89ms** total compilation time.

![http://i.imgur.com/r3CeDJP.png](http://i.imgur.com/r3CeDJP.png)

@halter73 @davidfowl