@halter73 2016-07-13
@SajayAntony Do you still have the application mentioned in https://github.com/aspnet/Hosting/issues/781 that was used in testing the impact of #426?