@halter73 2016-07-13
Now I want to know if this makes things faster on coreclr. The smaller trees and avoiding compilation for singletons could be helpful even there.

:shipit: