@pakrym 2016-07-13
Customers app

## Before:
![http://i.imgur.com/hXrDVoj.png](http://i.imgur.com/hXrDVoj.png)

## After
![http://i.imgur.com/D8BSRKB.png](http://i.imgur.com/D8BSRKB.png)